+++ index.html
@@ -529,6 +529,6 @@
     <!-- Confetti Canvas pour les célébrations -->
     <canvas id="confetti-canvas"></canvas>
 
-    <script src="app.js"></script>
+    <script type="module" src="app.js"></script>
 </body>
 </html>
